requirements.txt
langchain-openai==0.1.0
keybert==0.7.0
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
flask==3.0.0
flask-cors==4.0.0
